Move useHeight to its own file
