+++ src/hooks/useHeight/useHeight.ts
@@ -1,0 +1,18 @@
+import { useEffect, useState } from 'react';
+
+export default function useHeight() {
+  const [height, setHeight] = useState(window.innerHeight);
+
+  useEffect(() => {
+    const onResize = () => {
+      setHeight(window.innerHeight);
+    };
+
+    window.addEventListener('resize', onResize);
+    return () => {
+      window.removeEventListener('resize', onResize);
+    };
+  });
+
+  return height + 'px';
+}
